LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-backdrop-expected.txt

Anchor
Dialog

FAIL ::backdrop can use position-try-options assert_equals: expected "300px" but got "9999px"
